*** empty log message ***

+++ src/fixed_array.d
@@ -64,12 +64,12 @@ struct FixedArray(T, uint capacity_, bool borrowChecked = false)
             static if (T.sizeof == 1)
                 alias Length = ubyte; // pack length
             else
-                alias Length = size_t;
+                alias Length = uint;
         }
         else static if (capacity <= ushort.max)
         {
             static if (T.sizeof <= 2)
-                alias Length = ushort; // pack length
+                alias Length = uint; // pack length
             else
                 alias Length = size_t;
         }
